First attempt at refactoring tests
